+++ haskell/src/Stablecoin/Client/Contract.hs
@@ -2,21 +2,34 @@
 -- SPDX-License-Identifier: MIT
 
 module Stablecoin.Client.Contract
-  ( mkInitialStorage
-  , InitialStorageData(..)
+  ( InitialStorageData(..)
+  , mkInitialStorage
   ) where
 
-import Michelson.Text (MText)
 import Morley.Client (AddressOrAlias)
-import Tezos.Address (Address)
 
-import Lorentz.Contracts.Stablecoin (Expiry, Roles(..), Storage, Storage'(..), metadataMap)
+import Lorentz as L
+import Lorentz.Contracts.Spec.TZIP16Interface (MetadataMap)
+import Lorentz.Contracts.Stablecoin (Expiry, Roles(..), Storage, Storage'(..))
+import Stablecoin.Client.Parser (ContractMetadataOptions(..))
 
 type family ComputeRegistryAddressType a where
   ComputeRegistryAddressType Address = Address
   ComputeRegistryAddressType AddressOrAlias = Maybe AddressOrAlias
 
+type family ComputeContractMetadataStorageType a where
+  ComputeContractMetadataStorageType Address = MetadataMap BigMap
+  ComputeContractMetadataStorageType AddressOrAlias = ContractMetadataOptions
+
 -- | The data needed in order to create the stablecoin contract's initial storage.
+-- This is used in two slightly different contexts, and some field types require
+-- changes accordingly.
+--
+-- 1. As an input to the `deploy` function when the
+-- registry contract addresses only contains the configurations for metadata contracts.
+--
+-- 2. As an input to the function that creates the raw initial storage value
+-- for the contract. There these fields will contain resolved addresses.
 data InitialStorageData addr = InitialStorageData
   { isdMasterMinter :: addr
   , isdContractOwner :: addr
@@ -26,12 +39,13 @@ data InitialStorageData addr = InitialStorageData
   , isdTokenName :: MText
   , isdTokenDecimals :: Natural
   , isdTokenMetadataRegistry :: ComputeRegistryAddressType addr
+  , isdContractMetadataStorage :: ComputeContractMetadataStorageType addr
   , isdDefaultExpiry :: Expiry
   }
 
 -- | Construct the stablecoin contract's initial storage in order to deploy it.
 mkInitialStorage :: InitialStorageData Address -> Storage
-mkInitialStorage (InitialStorageData {..}) =
+mkInitialStorage InitialStorageData {..} =
   Storage
     { sDefaultExpiry = isdDefaultExpiry
     , sLedger =  mempty
@@ -48,6 +62,6 @@ mkInitialStorage (InitialStorageData {..}) =
         }
     , sTokenMetadataRegistry = isdTokenMetadataRegistry
     , sTransferlistContract = isdTransferlist
-    , sMetadata = metadataMap
+    , sMetadata = isdContractMetadataStorage
     , sTotalSupply = 0
     }