+++ haskell/src/Stablecoin/Client/Main.hs
@@ -49,7 +49,7 @@ mainProgram (ClientArgs _ globalOptions cmd) = do
       -- prefix option was passed in via the CLI.
       aliasAlreadyExists <- checkIfAliasExists (coerce contractAlias)
 
-      (opHash, contractAddr, metadataRegAddr) <- deploy user (coerce contractAlias) InitialStorageData
+      (opHash, contractAddr, metadataRegAddr, mCMetadataAddr) <- deploy user (coerce contractAlias) InitialStorageData
         { isdMasterMinter = dcoMasterMinter
         , isdContractOwner = dcoContractOwner
         , isdPauser = dcoPauser
@@ -59,12 +59,15 @@ mainProgram (ClientArgs _ globalOptions cmd) = do
         , isdTokenDecimals = dcoTokenDecimals
         , isdTokenMetadataRegistry = dcoTokenMetadataRegistry
         , isdDefaultExpiry = dcoDefaultExpiry
+        , isdContractMetadataStorage = dcoContractMetadata
         }
 
       putTextLn "Contract was successfully deployed."
       putTextLn $ "Operation hash: " <> pretty opHash
       putTextLn $ "Contract address: " <> pretty contractAddr
       putTextLn $ "Metadata registry contract address: " <> pretty metadataRegAddr
+      whenJust mCMetadataAddr $ \addr ->
+        putTextLn $ "Contract metadata contract address: " <> pretty addr
 
       let printAlias = putTextLn $ "Created alias '" <> pretty contractAlias <> "'."