+++ haskell/test/Lorentz/Contracts/Test/FA1_2.hs
@@ -16,9 +16,9 @@ import Lorentz.Contracts.Test.ApprovableLedger (AlSettings(..), approvableLedger
 import Michelson.Test.Integrational
 import Michelson.Typed (untypeValue)
 
-import Lorentz.Contracts.Stablecoin (Roles(..))
+import Lorentz.Contracts.Stablecoin (Roles(..), MetadataUri(..), metadataMap)
 import Lorentz.Contracts.StablecoinFA1_2
-  (Parameter, Storage(..), metadataMap, stablecoinFA1_2Contract)
+  (Parameter, Storage(..), metadataJSON, stablecoinFA1_2Contract)
 
 alOriginationFunction :: Address -> AlSettings -> IntegrationalScenarioM (TAddress Parameter)
 alOriginationFunction adminAddr (AlInitAddresses addrBalances) = do
@@ -38,7 +38,7 @@ alOriginationFunction adminAddr (AlInitAddresses addrBalances) = do
             }
         , sTransferlistContract = Nothing
         , sTotalSupply = sum $ snd <$> addrBalances
-        , sMetadata = metadataMap
+        , sMetadata = metadataMap (CurrentContract metadataJSON True)
         }
 
   TAddress @Parameter <$> originate stablecoinFA1_2Contract "" (untypeValue $ toVal storage) (toMutez 0)